requirements.txt
# Core
python-dateutil<2.8.1,>=2.1
netaddr>=0.7.11
sqlitedict>=1.6.0
cherrypy>=18.1.0
cherrypy-cors>=1.6
coloredlogs<=10.0
asyncio-throttle==0.1.1

# AWS Provider
botocore>=1.20.21
boto3>=1.9.210
policyuniverse>=1.3.2.0

# GCP Provider
grpcio>=1.18.0
## Cloud Libraries
google-cloud-container>=2.1.0
google-cloud-core>=0.29.1
google-cloud-iam>=0.1.0
google-cloud-logging>=2.2.0
google-cloud-monitoring==1.1.0
google-cloud-resource-manager>=0.28.3
google-cloud-storage>=1.13.2
google-cloud-kms==1.3.0
## API Client Libraries
google-api-python-client>=1.7.8
oauth2client>=4.1.3
## Necessary since API Client Libraries are not thread-safe
httplib2shim>=0.0.3

# Azure Provider
## core requirements
azure-cli-core==2.12.0
## for RBAC in AAD
azure-graphrbac==0.61.1
adal==1.2.4
PyJWT==1.7.1  # FIXME - this is currently required due to the MSAL dependency but should be removed once we switch to ADAL
## for resources
azure-mgmt-storage==7.1.0
azure-mgmt-monitor==0.5.2
azure-mgmt-sql==0.11.0
azure-mgmt-security==0.4.1
azure-mgmt-keyvault==1.1.0
azure-mgmt-network==2.5.1
azure-mgmt-redis==6.0.0
azure-mgmt-web==0.47.0
azure-mgmt-compute==12.0.0
azure-mgmt-authorization==0.60.0

# Aliyun / Alibaba Cloud Provider
aliyun-python-sdk-core>=2.13.4
aliyun-python-sdk-ram>=3.0.1
aliyun-python-sdk-ocs>=0.0.4
aliyun-python-sdk-sts>=3.0.1
aliyun-python-sdk-actiontrail>=2.0.0
aliyun-python-sdk-vpc>=3.0.5
aliyun-python-sdk-ecs>=4.16.10
aliyun-python-sdk-rds>=2.3.9
aliyun-python-sdk-kms>=2.6.0
oss2>=2.8.0

# Oracle Cloud Infrastructure Provider
oci>=2.2.4

# Gsuite
google-auth-httplib2>=0.0.4
google-auth-oauthlib>=0.4.1